pyproject.toml
[project]
name = "researcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Andre Ratzenberger", email = "andre.ratzenberger@whiteduck.de" }
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "arxiv>=2.3.1",
    "fastapi>=0.121.2",
    "litellm>=1.80.0",
    "networkx>=3.5",
    "pydantic-settings>=2.12.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "tinydb>=4.8.2",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]

[project.scripts]
researcher = "researcher:main"

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"
